+++ src/scripts/components/tBody/index.js
@@ -1,0 +1,1 @@
+export {createTableBody} from "./tBody"
