+++ src/scripts/components/tHeader/index.js
@@ -1,0 +1,1 @@
+export {createTableHeader} from "./tHeader"
